move runtime modules boostrap into RuntimePlugin
