+++ lib/MainTemplate.js
@@ -14,7 +14,6 @@ const {
 } = require("webpack-sources");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const Template = require("./Template");
-const { compareIds } = require("./util/comparators");
 
 /** @typedef {import("webpack-sources").ConcatSource} ConcatSource */
 /** @typedef {import("webpack-sources").Source} Source */
@@ -93,6 +92,8 @@ module.exports = class MainTemplate {
 			require: new SyncWaterfallHook(["source", "renderContext"]),
 			/** @type {SyncWaterfallHook<string, RenderBootstrapContext>} */
 			requireExtensions: new SyncWaterfallHook(["source", "renderContext"]),
+			/** @type {SyncWaterfallHook<string, RenderBootstrapContext>} */
+			runtime: new SyncWaterfallHook(["source", "renderContext"]),
 			/** @type {SyncWaterfallHook<string, Chunk, string>} */
 			beforeStartup: new SyncWaterfallHook(["source", "chunk", "hash"]),
 			/** @type {SyncWaterfallHook<string, RenderBootstrapContext>} */
@@ -132,19 +133,6 @@ module.exports = class MainTemplate {
 			(source, { chunk, hash, chunkGraph }) => {
 				/** @type {string[]} */
 				const buf = [];
-				if (chunkGraph.getNumberOfRuntimeModules(chunk) > 0) {
-					buf.push("// Bootstrap all runtime modules");
-					const runtimeModules = Array.from(
-						chunkGraph.getChunkRuntimeModulesIterable(chunk),
-						m => chunkGraph.getModuleId(m)
-					);
-					runtimeModules.sort(compareIds);
-					for (const moduleId of runtimeModules) {
-						buf.push(
-							`modules[${JSON.stringify(moduleId)}](0,0,__webpack_require__);`
-						);
-					}
-				}
 				if (chunkGraph.getNumberOfEntryModules(chunk) > 0) {
 					const runtimeRequirements = chunkGraph.getChunkRuntimeRequirements(
 						chunk
@@ -435,6 +423,7 @@ module.exports = class MainTemplate {
 			Template.asString(this.hooks.requireExtensions.call("", renderContext))
 		);
 		buf.push("");
+		buf.push(Template.asString(this.hooks.runtime.call("", renderContext)));
 		buf.push(Template.asString(this.hooks.beforeStartup.call("", chunk, hash)));
 		buf.push(Template.asString(this.hooks.startup.call("", renderContext)));
 		return buf;
