+++ lib/RuntimePlugin.js
@@ -6,7 +6,9 @@
 "use strict";
 
 const RuntimeGlobals = require("./RuntimeGlobals");
+const Template = require("./Template");
 const DefinePropertyGetterRuntimeModule = require("./runtime/DefinePropertyGetterRuntimeModule");
+const { compareIds } = require("./util/comparators");
 
 /** @typedef {import("./Chunk")} Chunk */
 /** @typedef {import("./Compiler")} Compiler */
@@ -68,6 +70,26 @@ class RuntimePlugin {
 					);
 					return true;
 				});
+			compilation.mainTemplate.hooks.runtime.tap(
+				"RuntimePlugin",
+				(source, { chunk, chunkGraph }) => {
+					const buf = [];
+					if (chunkGraph.getNumberOfRuntimeModules(chunk) > 0) {
+						buf.push("// Bootstrap all runtime modules");
+						const runtimeModules = Array.from(
+							chunkGraph.getChunkRuntimeModulesIterable(chunk),
+							m => chunkGraph.getModuleId(m)
+						);
+						runtimeModules.sort(compareIds);
+						for (const moduleId of runtimeModules) {
+							buf.push(
+								`modules[${JSON.stringify(moduleId)}](0,0,__webpack_require__);`
+							);
+						}
+					}
+					return Template.asString(buf);
+				}
+			);
 		});
 	}
 }
